llm/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.7.4,<3.0.0
tomli>=2.0.1
python-dotenv>=1.0.0
pyyaml>=6.0.1
httpx>=0.25.0
python-multipart>=0.0.6
loguru>=0.7.0
langchain-community>=0.2.17

# LLM dependencies
langchain==0.2.17
langchain-openai>=0.1.0

# Image generation dependencies
pillow>=10.0.0

# Development dependencies (optional)
# pytest>=7.4.2
# pytest-asyncio>=0.21.1
# black>=23.9.1
# isort>=5.12.0
# mypy>=1.5.1
# flake8>=6.1.0
